// types/post.ts
// Generated by https://quicktype.io

export interface Post {
  id: string;
  tags: string[];
  markup: string;
  publishedAt: number;
  mainImageAttributionLink: null;
  fromSlack: boolean;
  firstSeenAt: boolean;
  title: string;
  mainImage: string;
  slug: string;
  createdAt: number;
  wordCount: number;
  editor: Editor;
  excerpt: string;
  profile: PostProfile;
  images: Images;
  estimatedTime: number;
  published: boolean;
  reactions: Reactions;
  relatedStories: RelatedStory[];
}

export interface Editor {
  displayName: string;
  id: string;
  handle: string;
}

export interface Images {
  [x: string]: {
    width: number;
    height: number;
  };
}

export interface PostProfile {
  github: string;
  adText: string;
  preferredCharity: null;
  showStatsPublicly: boolean;
  publicBookmarks: boolean;
  id: string;
  bio: string;
  youtube: string;
  avatar: string;
  twitter: string;
  adLink: string;
  adIcon: string;
  displayName: string;
  handle: string;
}

export interface Reactions {
  sections: Sections;
  total: number;
  storyReactions: StoryReactions;
}

export interface Sections {}

export interface StoryReactions {
  light: number;
  boat: number;
  heart: number;
  money: number;
}

export interface RelatedStory {
  companyName?: string;
  image?: string;
  type?: string;
  active?: boolean;
  tags: string[];
  category?: Category[];
  website?: string;
  text?: string;
  link?: string;
  reactionsCount: number;
  id?: string;
  owner?: string;
  publishedAt?: number;
  slug?: string;
  title?: string;
  firstSeenAt?: string;
  excerpt?: string;
  estimatedTime?: number;
  createdAt?: number;
  profile?: RelatedStoryProfile;
  mainImage?: string;
  published?: boolean;
}

export interface Category {
  value: string;
  label: string;
}

export interface RelatedStoryProfile {
  youtube?: string;
  adLink?: string;
  id: string;
  displayName: string;
  github: string;
  twitter?: string;
  adText?: string;
  facebook?: string;
  paymentPointer?: string;
  linkedin: string;
  handle: string;
  avatar: string;
  bio: string;
}